CMT #46: Found structural flaw in the API
All the functions are working as required. But they aren't making changes to the central copy of the data. They are making copy to their own copy that is NOT getting reflected throughout the system.

Make use of pointers to share values among functions. Halting the filling of other network data until this problem is resolved.

Restructured some attributes in the test_data.